--- packages/react-native-client/src/common/eventEmitter.ts
@@ -1,21 +1,0 @@
-import { NativeModulesProxy, EventEmitter } from 'expo-modules-core';
-
-import RNFishjamClientModule from '../RNFishjamClientModule';
-
-export const ReceivableEvents = {
-  IsCameraOn: 'IsCameraOn',
-  IsMicrophoneOn: 'IsMicrophoneOn',
-  IsScreencastOn: 'IsScreencastOn',
-  SimulcastConfigUpdate: 'SimulcastConfigUpdate',
-  PeersUpdate: 'PeersUpdate',
-  AudioDeviceUpdate: 'AudioDeviceUpdate',
-  SendMediaEvent: 'SendMediaEvent',
-  BandwidthEstimation: 'BandwidthEstimation',
-  ReconnectionRetriesLimitReached: 'ReconnectionRetriesLimitReached',
-  ReconnectionStarted: 'ReconnectionStarted',
-  Reconnected: 'Reconnected',
-} as const;
-
-export const eventEmitter = new EventEmitter(
-  RNFishjamClientModule ?? NativeModulesProxy.RNFishjamClient,
-);
