+++ packages/react-native-client/src/hooks/useBandwidthEstimation.ts
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 
-import { ReceivableEvents } from '../common/eventEmitter';
-import { useFishjamEvent } from './useFishjamEvent';
+import { ReceivableEvents, useFishjamEvent } from './useFishjamEvent';
 
 export type BandwidthEstimationEvent = { BandwidthEstimation: number };
