+++ packages/react-native-client/src/hooks/useFishjamEvent.ts
@@ -1,5 +1,24 @@
 import { useEffect } from 'react';
-import { eventEmitter, ReceivableEvents } from '../common/eventEmitter';
+import { NativeModulesProxy, EventEmitter } from 'expo-modules-core';
+import RNFishjamClientModule from '../RNFishjamClientModule';
+
+export const ReceivableEvents = {
+  IsCameraOn: 'IsCameraOn',
+  IsMicrophoneOn: 'IsMicrophoneOn',
+  IsScreencastOn: 'IsScreencastOn',
+  SimulcastConfigUpdate: 'SimulcastConfigUpdate',
+  PeersUpdate: 'PeersUpdate',
+  AudioDeviceUpdate: 'AudioDeviceUpdate',
+  SendMediaEvent: 'SendMediaEvent',
+  BandwidthEstimation: 'BandwidthEstimation',
+  ReconnectionRetriesLimitReached: 'ReconnectionRetriesLimitReached',
+  ReconnectionStarted: 'ReconnectionStarted',
+  Reconnected: 'Reconnected',
+} as const;
+
+const eventEmitter = new EventEmitter(
+  RNFishjamClientModule ?? NativeModulesProxy.RNFishjamClient,
+);
 
 export function useFishjamEvent<Xd>(
   eventName: keyof typeof ReceivableEvents,
