+++ packages/react-native-client/src/hooks/useParticipants.ts
@@ -2,8 +2,7 @@ import { useCallback, useEffect, useState } from 'react';
 
 import { GenericMetadata, TrackEncoding, TrackMetadata } from '../types';
 import RNFishjamClientModule from '../RNFishjamClientModule';
-import { ReceivableEvents } from '../common/eventEmitter';
-import { useFishjamEvent } from './useFishjamEvent';
+import { ReceivableEvents, useFishjamEvent } from './useFishjamEvent';
 
 export type TrackType = 'Audio' | 'Video';
